Restructure card packages
