+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/EditActivity.java
@@ -41,10 +41,10 @@ import it.niedermann.nextcloud.deck.model.User;
 import it.niedermann.nextcloud.deck.model.full.FullCard;
 import it.niedermann.nextcloud.deck.persistence.sync.SyncManager;
 import it.niedermann.nextcloud.deck.ui.board.BoardAdapter;
-import it.niedermann.nextcloud.deck.ui.card.CardAttachmentsFragment;
-import it.niedermann.nextcloud.deck.ui.card.CardDetailsFragment;
+import it.niedermann.nextcloud.deck.ui.card.attachments.CardAttachmentsFragment;
+import it.niedermann.nextcloud.deck.ui.card.details.CardDetailsFragment;
 import it.niedermann.nextcloud.deck.ui.card.CardTabAdapter;
-import it.niedermann.nextcloud.deck.ui.card.CommentDialogFragment;
+import it.niedermann.nextcloud.deck.ui.card.activities.CommentDialogFragment;
 import it.niedermann.nextcloud.deck.ui.exception.ExceptionHandler;
 import it.niedermann.nextcloud.deck.util.CardUtil;
 import kotlin.NotImplementedError;
