--- app/src/main/java/it/niedermann/nextcloud/deck/ui/card/CardItemTouchHelper.java
@@ -1,29 +1,0 @@
-package it.niedermann.nextcloud.deck.ui.card;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.ItemTouchHelper;
-
-public class CardItemTouchHelper extends ItemTouchHelper {
-    public CardItemTouchHelper(CardAdapter adapter) {
-        super(new ItemTouchHelper.SimpleCallback(
-                ItemTouchHelper.UP |
-                        ItemTouchHelper.DOWN |
-                        ItemTouchHelper.LEFT |
-                        ItemTouchHelper.RIGHT |
-                        ItemTouchHelper.START |
-                        ItemTouchHelper.END,
-                0) {
-            @Override
-            public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, @NonNull RecyclerView.ViewHolder target) {
-                adapter.moveItem(viewHolder.getAdapterPosition(), target.getAdapterPosition());
-                return true;
-            }
-
-            @Override
-            public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int i) {
-                /* silence is gold. */
-            }
-        });
-    }
-}
