+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/card/LabelAutoCompleteAdapter.java
@@ -23,7 +23,7 @@ public class LabelAutoCompleteAdapter extends AutoCompleteAdapter<Label> {
     private String lastFilterText;
     private boolean canManage = false;
 
-    LabelAutoCompleteAdapter(@NonNull ComponentActivity activity, long accountId, long boardId, long cardId) {
+    public LabelAutoCompleteAdapter(@NonNull ComponentActivity activity, long accountId, long boardId, long cardId) {
         super(activity, accountId, boardId, cardId);
         syncManager.getFullBoardById(accountId, boardId).observe(activity, (fullBoard) -> {
             if (fullBoard.getBoard().isPermissionManage()) {
@@ -93,7 +93,7 @@ public class LabelAutoCompleteAdapter extends AutoCompleteAdapter<Label> {
         };
     }
 
-    String getLastFilterText() {
+    public String getLastFilterText() {
         return this.lastFilterText;
     }
 }
